Feat: cleanup
