+++ src/prime_rl/configs/rl.py
@@ -117,6 +117,22 @@ class SharedWeightBroadcastConfig(BaseConfig):
     port: Annotated[int, Field(description="The port to use for NCCL weight broadcast.")] = 29501
     timeout: Annotated[int, Field(description="The timeout in seconds for NCCL weight broadcast.")] = 1200
 
+    use_kernel_format_transfer: Annotated[
+        bool,
+        Field(
+            description="Transfer weights in vLLM kernel format instead of HF checkpoint format. "
+            "Avoids the HF conversion intermediate step and allows direct in-place weight updates."
+        ),
+    ] = False
+
+    quantize_fp8: Annotated[
+        bool,
+        Field(
+            description="Quantize weights to FP8 (e4m3) with block-wise scaling during kernel format transfer. "
+            "Only used when use_kernel_format_transfer is True."
+        ),
+    ] = False
+
 
 class BaseDeploymentConfig(BaseModel):
     """Configures a base deployment."""
@@ -280,6 +296,16 @@ class RLConfig(BaseConfig):
 
     deployment: DeploymentConfig = SingleNodeDeploymentConfig()
 
+    allow_different_inference_model: Annotated[
+        bool,
+        Field(
+            description="Allow the inference server to use a different model name than the trainer. "
+            "When enabled, the orchestrator uses the inference model name for querying. "
+            "Useful for kernel format weight transfer where the trainer uses a bf16 model "
+            "and inference uses a quantized (e.g. FP8) variant.",
+        ),
+    ] = False
+
     slurm: Annotated[SlurmConfig | None, Field(description="SLURM configuration. If None, will run locally.")] = None
 
     dry_run: Annotated[bool, Field(description="Only validate and dump resolved configs and exit early.")] = False
@@ -416,24 +442,22 @@ class RLConfig(BaseConfig):
     @model_validator(mode="after")
     def auto_setup_model(self):
         """Auto-setup shared model config for trainer, orchestrator, and inference."""
-        import os
-
-        kernel_transfer = os.environ.get("PRIME_RL_KERNEL_WEIGHT_TRANSFER", "0") == "1"
+        allow_different = self.allow_different_inference_model
 
         if self.model is not None:
             self.trainer.model.name = self.model.name
             if self.inference is not None:
                 default_name = type(self.inference.model).model_fields["name"].default
-                if not kernel_transfer or self.inference.model.name == default_name:
+                if not allow_different or self.inference.model.name == default_name:
                     self.inference.model.name = self.model.name
 
             # Orchestrator must use the inference model name so it queries the correct model
-            if kernel_transfer and self.inference is not None and self.inference.model.name != self.model.name:
+            if allow_different and self.inference is not None and self.inference.model.name != self.model.name:
                 self.orchestrator.model.name = self.inference.model.name
             else:
                 self.orchestrator.model.name = self.model.name
 
-        validate_shared_model_name(self.trainer, self.orchestrator, self.inference)
+        validate_shared_model_name(self.trainer, self.orchestrator, self.inference, allow_different=allow_different)
 
         return self
 
@@ -491,11 +515,14 @@ class RLConfig(BaseConfig):
                     inference_world_size=inference_world_size,
                     port=self.weight_broadcast.port,
                     timeout=self.weight_broadcast.timeout,
+                    use_kernel_format_transfer=self.weight_broadcast.use_kernel_format_transfer,
+                    quantize_fp8=self.weight_broadcast.quantize_fp8,
                 )
                 self.orchestrator.weight_broadcast = OrchestratorNCCLWeightBroadcastConfig(
                     type=self.weight_broadcast.type,
                     port=self.weight_broadcast.port,
                     timeout=self.weight_broadcast.timeout,
+                    use_kernel_format_transfer=self.weight_broadcast.use_kernel_format_transfer,
                 )
             elif self.weight_broadcast.type == "filesystem":
                 self.trainer.weight_broadcast = TrainerFileSystemWeightBroadcastConfig()
