+++ src/prime_rl/configs/trainer.py
@@ -595,6 +595,22 @@ class NCCLWeightBroadcastConfig(BaseWeightBroadcastConfig):
     # TODO: Should not be configurable, but auto-inferred
     inference_world_size: Annotated[int, Field(description="The number of GPUs used for inference.")] = 1
 
+    use_kernel_format_transfer: Annotated[
+        bool,
+        Field(
+            description="Transfer weights in vLLM kernel format instead of HF checkpoint format. "
+            "Avoids the HF conversion intermediate step and allows direct in-place weight updates."
+        ),
+    ] = False
+
+    quantize_fp8: Annotated[
+        bool,
+        Field(
+            description="Quantize weights to FP8 (e4m3) with block-wise scaling during kernel format transfer. "
+            "Only used when use_kernel_format_transfer is True."
+        ),
+    ] = False
+
 
 WeightBroadcastConfig: TypeAlias = Annotated[
     FileSystemWeightBroadcastConfig | NCCLWeightBroadcastConfig, Field(discriminator="type")
