+++ src/prime_rl/orchestrator/orchestrator.py
@@ -326,6 +326,7 @@ async def orchestrate(config: OrchestratorConfig):
             config.weight_broadcast.host,
             config.weight_broadcast.port,
             config.weight_broadcast.timeout,
+            use_kernel_format_transfer=config.weight_broadcast.use_kernel_format_transfer,
         )
 
     # Setup training batch sender for sending training examples to trainer
