+++ src/prime_rl/trainer/models/base.py
@@ -98,6 +98,27 @@ class PreTrainedModelPrimeRL(PreTrainedModel):
         """
         raise NotImplementedError(f"convert_layer_to_prime is not implemented for {cls.__name__}")
 
+    @classmethod
+    def convert_layer_to_vllm_kernel(
+        cls, state_dict: dict[str, Tensor], layer_idx: int, quantize_fp8: bool = False
+    ) -> dict[str, Tensor]:
+        """
+        Convert a single layer's state dict from PrimeRL format to vLLM kernel format.
+
+        This produces weights in the format vLLM uses at inference time (fused projections,
+        stacked experts, optional FP8 quantization), enabling direct in-place weight updates
+        via NCCL broadcast without going through the HF checkpoint format.
+
+        Args:
+            state_dict: The state dict containing the layer's weights in PrimeRL format.
+            layer_idx: The index of the layer to convert.
+            quantize_fp8: Whether to quantize weights to FP8 (e4m3) with block-wise scaling.
+
+        Returns:
+            A new dict with weights in vLLM kernel format.
+        """
+        raise NotImplementedError(f"convert_layer_to_vllm_kernel is not implemented for {cls.__name__}")
+
     def init_buffers_post_meta(self) -> None:
         """
         Initialize buffers that are not in the state dict after loading with meta device.
