+++ src/prime_rl/utils/client.py
@@ -301,7 +301,13 @@ async def unload_lora_adapter(admin_clients: list[AsyncClient], lora_name: str) 
     await asyncio.gather(*[_unload_lora_adapter(admin_client) for admin_client in admin_clients])
 
 
-async def init_nccl_broadcast(admin_clients: list[AsyncClient], host: str, port: int, timeout: int) -> None:
+async def init_nccl_broadcast(
+    admin_clients: list[AsyncClient],
+    host: str,
+    port: int,
+    timeout: int,
+    use_kernel_format_transfer: bool = False,
+) -> None:
     """Make a HTTP post request to the vLLM server to initialize the NCCL broadcast."""
     logger = get_logger()
 
@@ -317,6 +323,7 @@ async def init_nccl_broadcast(admin_clients: list[AsyncClient], host: str, port:
                     "server_rank": client_num,
                     "num_inference_server": len(admin_clients),
                     "timeout": timeout,
+                    "use_kernel_format_transfer": use_kernel_format_transfer,
                 },
             )
             response.raise_for_status()
